+++ src/org/schema/traits/PlaceTrait.php
@@ -8,7 +8,7 @@ use org\schema\creativeWork\Certification;
 use org\schema\creativeWork\Map;
 use org\schema\creativeWork\medias\ImageObject;
 use org\schema\creativeWork\Photograph;
-use org\schema\creativeWork\Website;
+use org\schema\creativeWork\WebSite;
 use org\schema\DefinedTerm;
 use org\schema\Demand;
 use org\schema\Event;